policy-monitor/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.42.0
jinja2==3.1.4
python-dotenv==1.0.1
